requirements.txt
customtkinter
pywifi
scapy
geocoder
requests
pillow
bleak
